requirements/development.txt
asgiref==3.4.1
attrs==21.2.0
Django==4.0
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.13.1
drf-spectacular==0.21.0
inflection==0.5.1
jsonschema==4.3.1
psycopg2-binary==2.9.2
pyrsistent==0.18.0
python-decouple==3.5
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
